PB-1897: Resolve cyclic import
