+++ .pylintrc
@@ -64,10 +64,7 @@ disable=missing-docstring,
     missing-module-docstring,
     useless-object-inheritance,
     unnecessary-lambda-assignment,
-    too-many-positional-arguments,
-    # there is currently a bug preventing disabling cyclic imports on a specific line
-    # see https://github.com/pylint-dev/pylint/issues/9490
-    cyclic-import
+    too-many-positional-arguments
 
 # Enable the message, report, category or checker with the given id(s). You can
 # either give multiple identifier separated by comma (,) or put this option
